Bug 886: Using integer instead of double so that we don't get floating points conversion problems
git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@5124 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ experiments/statsound/src/java/server/pt/utl/ist/elab/driver/statsound/StatSoundStampDriver.java
@@ -102,7 +102,7 @@ public class StatSoundStampDriver extends AbstractStampDriver {
 
 	private static final String JMF_PACKAGE = "com.linkare.rec.jmf";
 
-	private static final double SAMPLE_RATE = 11025;
+	private static final int SAMPLE_RATE = 11025;
 
 	private static final String CAPTURE_LOCATOR = "capture://" + SAMPLE_RATE + "/16/2";
